+++ test/envtest/controller.go
@@ -36,6 +36,10 @@ func NewManager(t *testing.T, ctx context.Context, cfg *rest.Config, s *runtime.
 			BindAddress: "0",
 		},
 		Controller: config.Controller{
+			// We don't want to hide panics in tests so expose them by failing the test run rather
+			// than silently ignoring them.
+			RecoverPanic: lo.ToPtr(false),
+
 			// This is needed because controller-runtime keeps a global list of controller
 			// names and panics if there are duplicates.
 			// This is a workaround for that in tests.